Hw04 auth

+++ model/index.js
@@ -1,36 +1,59 @@
 const Contact = require("./schemas/contact");
 
+const getAll = async (userId, query) => {
+  const {
+    limit = 5,
+    offset = 0,
+    sortBy,
+    sortByDesc,
+    filter, // name|age|isVaccinated
+    favorite = null,
+  } = query;
+  const optionsSearch = { owner: userId };
+  if (favorite !== null) {
+    optionsSearch.favorite = favorite;
+  }
+  const results = await Contact.paginate(optionsSearch, {
+    limit,
+    offset,
+    select: filter ? filter.split("|").join(" ") : "",
+    sort: {
+      ...(sortBy ? { [`${sortBy}`]: 1 } : {}),
+      ...(sortByDesc ? { [`${sortByDesc}`]: -1 } : {}),
+    },
+  });
+  const { docs: contacts, totalDocs: total } = results;
+  return { total, limit, offset, contacts };
+};
 
+const getContactById = async (userId, id) => {
+  const result = await Contact.findOne({ _id: id, owner: userId }).populate({
+    path: "owner",
+    select: "email subscription -_id",
+  });
+  return result;
+};
 
-const getAll = async () => {
-  const results = await Contact.find({})
-  return results
-}
-
-const getContactById = async (id) => {
-  const result = await Contact.findOne({ _id: id })
-  return result
-}
-
-const removeContact = async (id) => {
-  const result = await Contact.findByIdAndRemove({ _id: id })
-  return result
+const removeContact = async (userId, id) => {
+  const result = await Contact.findByIdAndRemove({ _id: id, owner: userId });
+  return result;
 };
 
 const addContact = async (body) => {
-  const result = await Contact.create(body)
-  return result
+  const result = await Contact.create(body);
+  return result;
 };
 
-const updateContact = async (id, body) => {
+const updateContact = async (userId, id, body) => {
   const result = await Contact.findOneAndUpdate(
     {
       _id: id,
+      owner: userId,
     },
     { ...body },
-    { new: true },
-  )
-  return result
+    { new: true }
+  );
+  return result;
 };
 
 module.exports = {
